source_code/netweaver/src/main/java/com/sap/netweaver/porta/core/ProxyException.java
/*******************************************************************************
 * Copyright (c) 2009, 2010 SAP AG.
 * All rights reserved. This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v1.0
 * which accompanies this distribution, and is available at
 * http://www.eclipse.org/legal/epl-v10.html
 *
 * Contributors:
 *    Kaloyan Raev (SAP AG) - initial API and implementation
 *******************************************************************************/
package com.sap.netweaver.porta.core;

public class ProxyException extends CoreException {
	
	private static final long serialVersionUID = 5275691210623500509L;
	
	private String host;
	
	public ProxyException(String host) {
		super("The host " + host + " cannot be resolved by the Axis runtime, but can be resolved in the local network. Perhaps, there is probably a proxy configured. The host should be configured to bypass the proxy settings. ");
		this.host = host;
	}
	
	public String getHost() {
		return host;
	}

}
